Update tests: Added support for bool arrays and wrapper functions
likely necessary to split lexer into modules
